domx: Reuse universal sstate across domains
The native tools and crosssdk packages are built separately for every
domain. Reuse of sstate increases the first build time because those
projects build once only.
There could be a more proper solution, but the variable for path prefix
SSTATE_EXTRAPATH is hardcoded in poky and the is no way to overwrite it
in user recipes.

Signed-off-by: Valerii Chubar <valerii_chubar@epam.com>